Finalisation des relations entre entités et création de la fonction d'inscription aux sorties

+++ site-AMCA-2023/src/Controller/PostsController.php
@@ -33,6 +33,8 @@ class PostsController extends AbstractController
             $entityManager->persist($post);
             $entityManager->flush();
 
+            $this->addFlash('success', 'Post créé avec succès');
+
             return $this->redirectToRoute('app_posts_index', [], Response::HTTP_SEE_OTHER);
         }
 
@@ -45,11 +47,20 @@ class PostsController extends AbstractController
     #[Route('/{id}/edit', name: 'app_posts_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Posts $post, EntityManagerInterface $entityManager): Response
     {
+        $user = $this->getUser();
+
+        // Check if the user is the administrator or the author of the post
+        if (!$this->isGranted('ROLE_ADMIN') && $post->getUser() !== $user) {
+            throw $this->createAccessDeniedException('Vous n\'avez pas la permission de modifier ce post.');
+        }
+
         $form = $this->createForm(PostsType::class, $post);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $entityManager->flush();
+
+            $this->addFlash('success', 'Post modifié avec succès');
 
             return $this->redirectToRoute('app_posts_index', [], Response::HTTP_SEE_OTHER);
         }
@@ -63,9 +74,17 @@ class PostsController extends AbstractController
     #[Route('/{id}', name: 'app_posts_delete', methods: ['POST'])]
     public function delete(Request $request, Posts $post, EntityManagerInterface $entityManager): Response
     {
+        $user = $this->getUser();
+        // Check if the user is the administrator or the author of the post
+        if (!$this->isGranted('ROLE_ADMIN') && $post->getUser() !== $user) {
+            throw $this->createAccessDeniedException('Vous n\'avez pas la permission de supprimer ce post.');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$post->getId(), $request->request->get('_token'))) {
             $entityManager->remove($post);
             $entityManager->flush();
+
+            $this->addFlash('success', 'Post supprimé avec succès');
         }
 
         return $this->redirectToRoute('app_posts_index', [], Response::HTTP_SEE_OTHER);
